+++ CHANGES.rst
@@ -16,6 +16,7 @@ Changelog
 
 **Fixed**
 
+- #132 Autofill Batch field in Sample Add form when Batch defined in request
 - #131 Autofill ClientPatientID field in Sample Add form when Patient defined
 
 **Security**